Split up code to runloop.c
