+++ griffin/griffin.c
@@ -1191,6 +1191,7 @@ GIT
 RETROARCH
 ============================================================ */
 #include "../retroarch.c"
+#include "../runloop.c"
 #include "../command.c"
 #include "../libretro-common/queues/task_queue.c"
