+++ retroarch.c
@@ -1,6 +1,6 @@
 /*  RetroArch - A frontend for libretro.
  *  Copyright (C) 2010-2014 - Hans-Kristian Arntzen
- *  Copyright (C) 2011-2017 - Daniel De Matteis
+ *  Copyright (C) 2011-2021 - Daniel De Matteis
  *  Copyright (C) 2012-2015 - Michael Lelli
  *  Copyright (C) 2014-2017 - Jean-André Santoni
  *  Copyright (C) 2016-2019 - Brad Parker
@@ -9200,14 +9200,12 @@ static bool path_init_subsystem(struct rarch_state *p_rarch)
    return true;
 }
 
-static void path_init_savefile(void)
+static void path_init_savefile(runloop_state_t *p_runloop)
 {
-   bool    should_sram_be_used = runloop_state.rarch_use_sram
-      && !runloop_state.rarch_is_sram_save_disabled;
+   bool    should_sram_be_used = p_runloop->rarch_use_sram
+      && !p_runloop->rarch_is_sram_save_disabled;
 
-   runloop_state.rarch_use_sram     = should_sram_be_used;
-
-   if (!runloop_state.rarch_use_sram)
+   if (!(p_runloop->rarch_use_sram = should_sram_be_used))
    {
       RARCH_LOG("[SRAM]: %s\n",
             msg_hash_to_str(MSG_SRAM_WILL_NOT_BE_SAVED));
@@ -10409,36 +10407,35 @@ bool menu_driver_is_alive(void)
 
 /* MESSAGE QUEUE */
 
-static void retroarch_msg_queue_deinit(void)
+static void runloop_msg_queue_deinit(runloop_state_t *p_runloop)
 {
    RUNLOOP_MSG_QUEUE_LOCK(runloop_state);
 
-   msg_queue_deinitialize(&runloop_state.msg_queue);
+   msg_queue_deinitialize(&p_runloop->msg_queue);
 
    RUNLOOP_MSG_QUEUE_UNLOCK(runloop_state);
 #ifdef HAVE_THREADS
-   slock_free(runloop_state.msg_queue_lock);
-   runloop_state.msg_queue_lock = NULL;
+   slock_free(p_runloop->msg_queue_lock);
+   p_runloop->msg_queue_lock = NULL;
 #endif
 
-   runloop_state.msg_queue_size = 0;
+   p_runloop->msg_queue_size = 0;
 }
 
-static void retroarch_msg_queue_init(void)
+static void runloop_msg_queue_init(runloop_state_t *p_runloop)
 {
-   retroarch_msg_queue_deinit();
-   msg_queue_initialize(&runloop_state.msg_queue, 8);
+   runloop_msg_queue_deinit(p_runloop);
+   msg_queue_initialize(&p_runloop->msg_queue, 8);
 
 #ifdef HAVE_THREADS
-   runloop_state.msg_queue_lock   = slock_new();
+   p_runloop->msg_queue_lock   = slock_new();
 #endif
 }
 
 #ifdef HAVE_THREADS
-static void retroarch_autosave_deinit(void)
+static void runloop_autosave_deinit(runloop_state_t *p_runloop)
 {
-   const bool rarch_use_sram   = runloop_state.rarch_use_sram;
-   if (rarch_use_sram)
+   if (p_runloop->rarch_use_sram)
       autosave_deinit();
 }
 #endif
@@ -12027,7 +12024,7 @@ static bool command_event_disk_control_append_image(
       return false;
 
 #ifdef HAVE_THREADS
-   retroarch_autosave_deinit();
+   runloop_autosave_deinit(&runloop_state);
 #endif
 
    /* TODO/FIXME: Need to figure out what to do with subsystems case. */
@@ -13888,7 +13885,7 @@ bool command_event(enum event_command cmd, void *data)
          break;
       case CMD_EVENT_AUTOSAVE_INIT:
 #ifdef HAVE_THREADS
-         retroarch_autosave_deinit();
+         runloop_autosave_deinit(&runloop_state);
          {
 #ifdef HAVE_NETWORKING
             unsigned autosave_interval =
@@ -15368,7 +15365,7 @@ void main_exit(void *args)
    runloop_state.rarch_block_config_read = false;
 #endif
 
-   retroarch_msg_queue_deinit();
+   runloop_msg_queue_deinit(&runloop_state);
    driver_uninit(p_rarch, DRIVERS_CMD_ALL);
 
    retro_main_log_file_deinit();
@@ -15469,7 +15466,7 @@ int rarch_main(int argc, char *argv[], void *data)
          input_config_set_device(i, RETRO_DEVICE_JOYPAD);
    }
 
-   retroarch_msg_queue_init();
+   runloop_msg_queue_init(&runloop_state);
 
    if (p_rarch->current_frontend_ctx)
    {
@@ -15583,439 +15580,6 @@ int main(int argc, char *argv[])
 #endif
 
 /* CORE OPTIONS */
-static const char *core_option_manager_parse_value_label(
-      const char *value, const char *value_label)
-{
-   /* 'value_label' may be NULL */
-   const char *label = string_is_empty(value_label) ?
-         value : value_label;
-
-   if (string_is_empty(label))
-      return NULL;
-
-   /* Any label starting with a digit (or +/-)
-    * cannot be a boolean string, and requires
-    * no further processing */
-   if (ISDIGIT((unsigned char)*label) ||
-       (*label == '+') ||
-       (*label == '-'))
-      return label;
-
-   /* Core devs have a habit of using arbitrary
-    * strings to label boolean values (i.e. enabled,
-    * Enabled, on, On, ON, true, True, TRUE, disabled,
-    * Disabled, off, Off, OFF, false, False, FALSE).
-    * These should all be converted to standard ON/OFF
-    * strings
-    * > Note: We require some duplication here
-    *   (e.g. MENU_ENUM_LABEL_ENABLED *and*
-    *    MENU_ENUM_LABEL_VALUE_ENABLED) in order
-    *   to match both localised and non-localised
-    *   strings. This function is not performance
-    *   critical, so these extra comparisons do
-    *   no harm */
-   if (string_is_equal_noncase(label, msg_hash_to_str(MENU_ENUM_LABEL_ENABLED)) ||
-       string_is_equal_noncase(label, msg_hash_to_str(MENU_ENUM_LABEL_VALUE_ENABLED)) ||
-       string_is_equal_noncase(label, "enable") ||
-       string_is_equal_noncase(label, "on") ||
-       string_is_equal_noncase(label, msg_hash_to_str(MENU_ENUM_LABEL_VALUE_ON)) ||
-       string_is_equal_noncase(label, "true") ||
-       string_is_equal_noncase(label, msg_hash_to_str(MENU_ENUM_LABEL_VALUE_TRUE)))
-      label = msg_hash_to_str(MENU_ENUM_LABEL_VALUE_ON);
-   else if (string_is_equal_noncase(label, msg_hash_to_str(MENU_ENUM_LABEL_DISABLED)) ||
-            string_is_equal_noncase(label, msg_hash_to_str(MENU_ENUM_LABEL_VALUE_DISABLED)) ||
-            string_is_equal_noncase(label, "disable") ||
-            string_is_equal_noncase(label, "off") ||
-            string_is_equal_noncase(label, msg_hash_to_str(MENU_ENUM_LABEL_VALUE_OFF)) ||
-            string_is_equal_noncase(label, "false") ||
-            string_is_equal_noncase(label, msg_hash_to_str(MENU_ENUM_LABEL_VALUE_FALSE)))
-      label = msg_hash_to_str(MENU_ENUM_LABEL_VALUE_OFF);
-
-   return label;
-}
-
-static bool core_option_manager_parse_variable(
-      core_option_manager_t *opt, size_t idx,
-      const struct retro_variable *var,
-      config_file_t *config_src)
-{
-   size_t i;
-   union string_list_elem_attr attr;
-   const char *val_start      = NULL;
-   char *value                = NULL;
-   char *desc_end             = NULL;
-   struct core_option *option = (struct core_option*)&opt->opts[idx];
-   struct config_entry_list
-      *entry                  = NULL;
-
-   /* All options are visible by default */
-   option->visible            = true;
-
-   if (!string_is_empty(var->key))
-      option->key             = strdup(var->key);
-   if (!string_is_empty(var->value))
-      value                   = strdup(var->value);
-
-   if (!string_is_empty(value))
-      desc_end                = strstr(value, "; ");
-
-   if (!desc_end)
-      goto error;
-
-   *desc_end    = '\0';
-
-   if (!string_is_empty(value))
-      option->desc    = strdup(value);
-
-   val_start          = desc_end + 2;
-   option->vals       = string_split(val_start, "|");
-
-   if (!option->vals)
-      goto error;
-
-   /* Legacy core option interface has no concept
-    * of value labels
-    * > Use actual values for display purposes */
-   attr.i             = 0;
-   option->val_labels = string_list_new();
-
-   if (!option->val_labels)
-      goto error;
-
-   /* > Loop over values and 'extract' labels */
-   for (i = 0; i < option->vals->size; i++)
-   {
-      const char *value       = option->vals->elems[i].data;
-      const char *value_label = core_option_manager_parse_value_label(
-            value, NULL);
-
-      /* Redundant safely check... */
-      value_label = string_is_empty(value_label) ?
-            value : value_label;
-
-      /* Append value label string */
-      string_list_append(option->val_labels, value_label, attr);
-   }
-
-   /* Legacy core option interface always uses first
-    * defined value as the default */
-   option->default_index = 0;
-   option->index         = 0;
-
-   if (config_src)
-      entry              = config_get_entry(config_src, option->key);
-   else
-      entry              = config_get_entry(opt->conf,  option->key);
-
-   /* Set current config value */
-   if (entry && !string_is_empty(entry->value))
-   {
-      for (i = 0; i < option->vals->size; i++)
-      {
-         if (string_is_equal(option->vals->elems[i].data, entry->value))
-         {
-            option->index = i;
-            break;
-         }
-      }
-   }
-
-   free(value);
-
-   return true;
-
-error:
-   free(value);
-   return false;
-}
-
-static bool core_option_manager_parse_option(
-      core_option_manager_t *opt, size_t idx,
-      const struct retro_core_option_definition *option_def,
-      config_file_t *config_src)
-{
-   size_t i;
-   union string_list_elem_attr attr;
-   struct config_entry_list
-      *entry                  = NULL;
-   size_t num_vals            = 0;
-   struct core_option *option = (struct core_option*)&opt->opts[idx];
-   const struct retro_core_option_value
-      *values                 = option_def->values;
-
-   /* All options are visible by default */
-   option->visible            = true;
-
-   if (!string_is_empty(option_def->key))
-      option->key             = strdup(option_def->key);
-
-   if (!string_is_empty(option_def->desc))
-      option->desc            = strdup(option_def->desc);
-
-   if (!string_is_empty(option_def->info))
-      option->info            = strdup(option_def->info);
-
-   /* Get number of values */
-   for (;;)
-   {
-      if (string_is_empty(values[num_vals].value))
-         break;
-      num_vals++;
-   }
-
-   if (num_vals < 1)
-      return false;
-
-   /* Initialise string lists */
-   attr.i             = 0;
-   option->vals       = string_list_new();
-   option->val_labels = string_list_new();
-
-   if (!option->vals || !option->val_labels)
-      return false;
-
-   /* Initialise default value */
-   option->default_index = 0;
-   option->index         = 0;
-
-   /* Extract value/label pairs */
-   for (i = 0; i < num_vals; i++)
-   {
-      const char *value       = values[i].value;
-      const char *value_label = values[i].label;
-
-      /* Append value string
-       * > We know that 'value' is always valid */
-      string_list_append(option->vals, value, attr);
-
-      /* Value label requires additional processing */
-      value_label = core_option_manager_parse_value_label(
-            value, value_label);
-
-      /* > Redundant safely check... */
-      value_label = string_is_empty(value_label) ?
-            value : value_label;
-
-      /* Append value label string */
-      string_list_append(option->val_labels, value_label, attr);
-
-      /* Check whether this value is the default setting */
-      if (!string_is_empty(option_def->default_value))
-      {
-         if (string_is_equal(option_def->default_value, value))
-         {
-            option->default_index = i;
-            option->index         = i;
-         }
-      }
-   }
-
-   if (config_src)
-      entry              = config_get_entry(config_src, option->key);
-   else
-      entry              = config_get_entry(opt->conf,  option->key);
-
-   /* Set current config value */
-   if (entry && !string_is_empty(entry->value))
-   {
-      for (i = 0; i < option->vals->size; i++)
-      {
-         if (string_is_equal(option->vals->elems[i].data, entry->value))
-         {
-            option->index = i;
-            break;
-         }
-      }
-   }
-
-   return true;
-}
-
-/**
- * core_option_manager_free:
- * @opt              : options manager handle
- *
- * Frees core option manager handle.
- **/
-static void core_option_manager_free(core_option_manager_t *opt)
-{
-   size_t i;
-
-   if (!opt)
-      return;
-
-   for (i = 0; i < opt->size; i++)
-   {
-      if (opt->opts[i].desc)
-         free(opt->opts[i].desc);
-      if (opt->opts[i].info)
-         free(opt->opts[i].info);
-      if (opt->opts[i].key)
-         free(opt->opts[i].key);
-
-      if (opt->opts[i].vals)
-         string_list_free(opt->opts[i].vals);
-      if (opt->opts[i].val_labels)
-         string_list_free(opt->opts[i].val_labels);
-
-      opt->opts[i].desc = NULL;
-      opt->opts[i].info = NULL;
-      opt->opts[i].key  = NULL;
-      opt->opts[i].vals = NULL;
-   }
-
-   if (opt->conf)
-      config_file_free(opt->conf);
-   free(opt->opts);
-   free(opt);
-}
-
-/**
- * core_option_manager_new_vars:
- * @conf_path        : Filesystem path to write core option config file to.
- * @src_conf_path    : Filesystem path from which to load initial config settings.
- * @vars             : Pointer to variable array handle.
- *
- * Legacy version of core_option_manager_new().
- * Creates and initializes a core manager handle.
- *
- * Returns: handle to new core manager handle, otherwise NULL.
- **/
-static core_option_manager_t *core_option_manager_new_vars(
-      const char *conf_path, const char *src_conf_path,
-      const struct retro_variable *vars)
-{
-   const struct retro_variable *var;
-   size_t size                       = 0;
-   config_file_t *config_src         = NULL;
-   core_option_manager_t *opt        = (core_option_manager_t*)
-      malloc(sizeof(*opt));
-
-   if (!opt)
-      return NULL;
-
-   opt->conf                         = NULL;
-   opt->conf_path[0]                 = '\0';
-   opt->opts                         = NULL;
-   opt->size                         = 0;
-   opt->updated                      = false;
-
-   if (!string_is_empty(conf_path))
-      if (!(opt->conf = config_file_new_from_path_to_string(conf_path)))
-         if (!(opt->conf = config_file_new_alloc()))
-            goto error;
-
-   strlcpy(opt->conf_path, conf_path, sizeof(opt->conf_path));
-
-   /* Load source config file, if required */
-   if (!string_is_empty(src_conf_path))
-      config_src = config_file_new_from_path_to_string(src_conf_path);
-
-   for (var = vars; var->key && var->value; var++)
-      size++;
-
-   if (size == 0)
-      goto error;
-
-   opt->opts = (struct core_option*)calloc(size, sizeof(*opt->opts));
-   if (!opt->opts)
-      goto error;
-
-   opt->size = size;
-   size      = 0;
-
-   for (var = vars; var->key && var->value; size++, var++)
-   {
-      if (!core_option_manager_parse_variable(opt, size, var, config_src))
-         goto error;
-   }
-
-   if (config_src)
-      config_file_free(config_src);
-
-   return opt;
-
-error:
-   if (config_src)
-      config_file_free(config_src);
-   core_option_manager_free(opt);
-   return NULL;
-}
-
-/**
- * core_option_manager_new:
- * @conf_path        : Filesystem path to write core option config file to.
- * @src_conf_path    : Filesystem path from which to load initial config settings.
- * @option_defs      : Pointer to variable array handle.
- *
- * Creates and initializes a core manager handle.
- *
- * Returns: handle to new core manager handle, otherwise NULL.
- **/
-static core_option_manager_t *core_option_manager_new(
-      const char *conf_path, const char *src_conf_path,
-      const struct retro_core_option_definition *option_defs)
-{
-   const struct retro_core_option_definition *option_def;
-   size_t size                       = 0;
-   config_file_t *config_src         = NULL;
-   core_option_manager_t *opt        = (core_option_manager_t*)
-      malloc(sizeof(*opt));
-
-   if (!opt)
-      return NULL;
-
-   opt->conf                         = NULL;
-   opt->conf_path[0]                 = '\0';
-   opt->opts                         = NULL;
-   opt->size                         = 0;
-   opt->updated                      = false;
-
-   if (!string_is_empty(conf_path))
-      if (!(opt->conf = config_file_new_from_path_to_string(conf_path)))
-         if (!(opt->conf = config_file_new_alloc()))
-            goto error;
-
-   strlcpy(opt->conf_path, conf_path, sizeof(opt->conf_path));
-
-   /* Load source config file, if required */
-   if (!string_is_empty(src_conf_path))
-      config_src = config_file_new_from_path_to_string(src_conf_path);
-
-   /* Note: 'option_def->info == NULL' is valid */
-   for (option_def = option_defs;
-        option_def->key && option_def->desc && option_def->values[0].value;
-        option_def++)
-      size++;
-
-   if (size == 0)
-      goto error;
-
-   opt->opts = (struct core_option*)calloc(size, sizeof(*opt->opts));
-   if (!opt->opts)
-      goto error;
-
-   opt->size = size;
-   size      = 0;
-
-   /* Note: 'option_def->info == NULL' is valid */
-   for (option_def = option_defs;
-        option_def->key && option_def->desc && option_def->values[0].value;
-        size++, option_def++)
-      if (!core_option_manager_parse_option(opt, size, option_def, config_src))
-         goto error;
-
-   if (config_src)
-      config_file_free(config_src);
-
-   return opt;
-
-error:
-   if (config_src)
-      config_file_free(config_src);
-   core_option_manager_free(opt);
-   return NULL;
-}
 
 /**
  * core_option_manager_flush:
@@ -16023,7 +15587,7 @@ error:
  *
  * Writes core option key-pair values to file.
  **/
-static void core_option_manager_flush(
+void core_option_manager_flush(
       config_file_t *conf,
       core_option_manager_t *opt)
 {
@@ -34143,40 +33707,6 @@ force_input_dirty:
 }
 #endif
 
-static retro_time_t rarch_core_runtime_tick(
-      float slowmotion_ratio,
-      retro_time_t current_time)
-{
-   retro_time_t frame_time              =
-      (1.0 / runloop_state.av_info.timing.fps) * 1000000;
-   bool runloop_slowmotion              = runloop_state.slowmotion;
-   bool runloop_fastmotion              = runloop_state.fastmotion;
-
-   /* Account for slow motion */
-   if (runloop_slowmotion)
-      return (retro_time_t)((double)frame_time * slowmotion_ratio);
-
-   /* Account for fast forward */
-   if (runloop_fastmotion)
-   {
-      /* Doing it this way means we miss the first frame after
-       * turning fast forward on, but it saves the overhead of
-       * having to do:
-       *    retro_time_t current_usec = cpu_features_get_time_usec();
-       *    libretro_core_runtime_last = current_usec;
-       * every frame when fast forward is off. */
-      retro_time_t current_usec                = current_time;
-      retro_time_t potential_frame_time        = current_usec -
-         runloop_state.libretro_core_runtime_last;
-      runloop_state.libretro_core_runtime_last = current_usec;
-
-      if (potential_frame_time < frame_time)
-         return potential_frame_time;
-   }
-
-   return frame_time;
-}
-
 static void retroarch_print_features(void)
 {
    char buf[2048];
@@ -35147,7 +34677,7 @@ static bool retroarch_parse_input_and_config(
    return verbosity_enabled;
 }
 
-static bool retroarch_validate_per_core_options(char *s,
+bool retroarch_validate_per_core_options(char *s,
       size_t len, bool mkdir,
       const char *core_name, const char *game_name)
 {
@@ -35183,7 +34713,7 @@ static bool retroarch_validate_per_core_options(char *s,
    return true;
 }
 
-static bool retroarch_validate_game_options(
+bool retroarch_validate_game_options(
       char *s, size_t len, bool mkdir)
 {
    const char *core_name       = runloop_state.system.info.library_name;
@@ -35193,7 +34723,7 @@ static bool retroarch_validate_game_options(
          core_name, game_name);
 }
 
-static bool retroarch_validate_folder_options(
+bool retroarch_validate_folder_options(
       char *s, size_t len, bool mkdir)
 {
    char folder_name[PATH_MAX_LENGTH];
@@ -35557,7 +35087,7 @@ bool retroarch_main_init(int argc, char *argv[])
    if (!string_is_empty(global->record.path))
       command_event(CMD_EVENT_RECORD_INIT, NULL);
 
-   path_init_savefile();
+   path_init_savefile(&runloop_state);
 
    command_event(CMD_EVENT_SET_PER_GAME_RESOLUTION, NULL);
 
@@ -35873,56 +35403,6 @@ void retroarch_menu_running_finished(bool quit)
 #endif
 }
 
-/**
- * rarch_game_specific_options:
- *
- * Returns: true (1) if a game specific core
- * options path has been found,
- * otherwise false (0).
- **/
-static bool rarch_game_specific_options(char **output)
-{
-   char game_options_path[PATH_MAX_LENGTH];
-   game_options_path[0] ='\0';
-
-   if (!retroarch_validate_game_options(
-            game_options_path,
-            sizeof(game_options_path), false) ||
-       !path_is_valid(game_options_path))
-      return false;
-
-   RARCH_LOG("%s %s\n",
-         msg_hash_to_str(MSG_GAME_SPECIFIC_CORE_OPTIONS_FOUND_AT),
-         game_options_path);
-   *output = strdup(game_options_path);
-   return true;
-}
-
-/**
- * rarch_folder_specific_options:
- *
- * Returns: true (1) if a folder specific core
- * options path has been found,
- * otherwise false (0).
- **/
-static bool rarch_folder_specific_options(char **output)
-{
-   char folder_options_path[PATH_MAX_LENGTH];
-   folder_options_path[0] ='\0';
-
-   if (!retroarch_validate_folder_options(
-            folder_options_path,
-            sizeof(folder_options_path), false) ||
-       !path_is_valid(folder_options_path))
-      return false;
-
-   RARCH_LOG("%s %s\n",
-         msg_hash_to_str(MSG_FOLDER_SPECIFIC_CORE_OPTIONS_FOUND_AT),
-         folder_options_path);
-   *output = strdup(folder_options_path);
-   return true;
-}
-
 static void runloop_task_msg_queue_push(
       retro_task_t *task, const char *msg,
       unsigned prio, unsigned duration,
@@ -35972,142 +35452,6 @@ static void runloop_task_msg_queue_push(
    else
 #endif
       runloop_msg_queue_push(msg, prio, duration, flush, NULL, MESSAGE_QUEUE_ICON_DEFAULT, MESSAGE_QUEUE_CATEGORY_INFO);
-}
-
-/* Fetches core options path for current core/content
- * - path: path from which options should be read
- *   from/saved to
- * - src_path: in the event that 'path' file does not
- *   yet exist, provides source path from which initial
- *   options should be extracted
- *
- *   NOTE: caller must ensure 
- *   path and src_path are NULL-terminated
- *  */
-static void runloop_init_core_options_path(
-      runloop_state_t *p_runloop,
-      settings_t *settings,
-      char *path, size_t len,
-      char *src_path, size_t src_len)
-{
-   char *game_options_path        = NULL;
-   char *folder_options_path      = NULL;
-   bool game_specific_options     = settings->bools.game_specific_options;
-
-   /* Check whether game-specific options exist */
-   if (game_specific_options &&
-       rarch_game_specific_options(&game_options_path))
-   {
-      /* Notify system that we have a valid core options
-       * override */
-      path_set(RARCH_PATH_CORE_OPTIONS, game_options_path);
-      p_runloop->game_options_active   = true;
-      p_runloop->folder_options_active = false;
-
-      /* Copy options path */
-      strlcpy(path, game_options_path, len);
-
-      free(game_options_path);
-   }
-   /* Check whether folder-specific options exist */
-   else if (game_specific_options &&
-            rarch_folder_specific_options(&folder_options_path))
-   {
-      /* Notify system that we have a valid core options
-       * override */
-      path_set(RARCH_PATH_CORE_OPTIONS, folder_options_path);
-      p_runloop->game_options_active   = false;
-      p_runloop->folder_options_active = true;
-
-      /* Copy options path */
-      strlcpy(path, folder_options_path, len);
-
-      free(folder_options_path);
-   }
-   else
-   {
-      char global_options_path[PATH_MAX_LENGTH];
-      char per_core_options_path[PATH_MAX_LENGTH];
-      bool per_core_options_exist   = false;
-      bool per_core_options         = !settings->bools.global_core_options;
-      const char *path_core_options = settings->paths.path_core_options;
-
-      global_options_path[0]        = '\0';
-      per_core_options_path[0]      = '\0';
-
-      if (per_core_options)
-      {
-         const char *core_name      = p_runloop->system.info.library_name;
-         /* Get core-specific options path
-          * > if retroarch_validate_per_core_options() returns
-          *   false, then per-core options are disabled (due to
-          *   unknown system errors...) */
-         per_core_options = retroarch_validate_per_core_options(
-               per_core_options_path, sizeof(per_core_options_path), true,
-               core_name, core_name);
-
-         /* If we can use per-core options, check whether an options
-          * file already exists */
-         if (per_core_options)
-            per_core_options_exist = path_is_valid(per_core_options_path);
-      }
-
-      /* If not using per-core options, or if a per-core options
-       * file does not yet exist, must fetch 'global' options path */
-      if (!per_core_options || !per_core_options_exist)
-      {
-         const char *options_path   = path_core_options;
-
-         if (!string_is_empty(options_path))
-            strlcpy(global_options_path,
-                  options_path, sizeof(global_options_path));
-         else if (!path_is_empty(RARCH_PATH_CONFIG))
-            fill_pathname_resolve_relative(
-                  global_options_path, path_get(RARCH_PATH_CONFIG),
-                  FILE_PATH_CORE_OPTIONS_CONFIG, sizeof(global_options_path));
-      }
-
-      /* Allocate correct path/src_path strings */
-      if (per_core_options)
-      {
-         strlcpy(path, per_core_options_path, len);
-
-         if (!per_core_options_exist)
-            strlcpy(src_path, global_options_path, src_len);
-      }
-      else
-         strlcpy(path, global_options_path, len);
-
-      /* Notify system that we *do not* have a valid core options
-       * options override */
-      p_runloop->game_options_active   = false;
-      p_runloop->folder_options_active = false;
-   }
-}
-
-static core_option_manager_t *runloop_init_core_options(
-      runloop_state_t *p_runloop,
-      settings_t *settings,
-      const struct retro_core_option_definition *option_defs)
-{
-   char options_path[PATH_MAX_LENGTH];
-   char src_options_path[PATH_MAX_LENGTH];
-
-   /* Ensure these are NULL-terminated */
-   options_path[0]                = '\0';
-   src_options_path[0]            = '\0';
-
-   /* Get core options file path */
-   runloop_init_core_options_path(
-         p_runloop,
-         settings,
-         options_path, sizeof(options_path),
-         src_options_path, sizeof(src_options_path));
-
-   if (!string_is_empty(options_path))
-      return core_option_manager_new(
-            options_path, src_options_path, option_defs);
-   return NULL;
 }
 
 void retroarch_init_task_queue(void)
@@ -36196,7 +35540,7 @@ bool rarch_ctl(enum rarch_ctl_state state, void *data)
          input_mapper_reset(&p_rarch->input_driver_mapper);
 
 #ifdef HAVE_THREADS
-         retroarch_autosave_deinit();
+         runloop_autosave_deinit(&runloop_state);
 #endif
 
          command_event(CMD_EVENT_RECORD_DEINIT, NULL);
@@ -36380,86 +35724,6 @@ bool rarch_ctl(enum rarch_ctl_state state, void *data)
    }
 
    return true;
-}
-
-static void runloop_deinit_core_options(
-      runloop_state_t *p_runloop,
-      const char *path_core_options)
-{
-   /* Check whether game-specific options file is being used */
-   if (!string_is_empty(path_core_options))
-   {
-      config_file_t *conf_tmp = NULL;
-
-      /* We only need to save configuration settings for
-       * the current core
-       * > If game-specific options file exists, have
-       *   to read it (to ensure file only gets written
-       *   if config values change)
-       * > Otherwise, create a new, empty config_file_t
-       *   object */
-      if (path_is_valid(path_core_options))
-         conf_tmp = config_file_new_from_path_to_string(path_core_options);
-
-      if (!conf_tmp)
-         conf_tmp = config_file_new_alloc();
-
-      if (conf_tmp)
-      {
-         core_option_manager_flush(
-               conf_tmp,
-               p_runloop->core_options);
-         RARCH_LOG("[Core Options]: Saved %s-specific core options to \"%s\"\n",
-               p_runloop->game_options_active 
-               ? "game" 
-               : "folder",
-               path_core_options);
-         config_file_write(conf_tmp, path_core_options, true);
-         config_file_free(conf_tmp);
-         conf_tmp = NULL;
-      }
-      path_clear(RARCH_PATH_CORE_OPTIONS);
-   }
-   else
-   {
-      const char *path = p_runloop->core_options->conf_path;
-      core_option_manager_flush(
-            p_runloop->core_options->conf,
-            p_runloop->core_options);
-      RARCH_LOG("[Core Options]: Saved core options file to \"%s\"\n", path);
-      config_file_write(p_runloop->core_options->conf, path, true);
-   }
-
-   p_runloop->game_options_active   = false;
-   p_runloop->folder_options_active = false;
-
-   if (p_runloop->core_options)
-      core_option_manager_free(p_runloop->core_options);
-   p_runloop->core_options          = NULL;
-}
-
-static core_option_manager_t *runloop_init_core_variables(
-      runloop_state_t *p_runloop,
-      settings_t *settings,
-      const struct retro_variable *vars)
-{
-   char options_path[PATH_MAX_LENGTH];
-   char src_options_path[PATH_MAX_LENGTH];
-
-   /* Ensure these are NULL-terminated */
-   options_path[0]     = '\0';
-   src_options_path[0] = '\0';
-
-   /* Get core options file path */
-   runloop_init_core_options_path(
-         p_runloop,
-         settings,
-         options_path, sizeof(options_path),
-         src_options_path, sizeof(src_options_path));
-
-   if (!string_is_empty(options_path))
-      return core_option_manager_new_vars(options_path, src_options_path, vars);
-   return NULL;
 }
 
 bool retroarch_is_forced_fullscreen(void)
@@ -36920,13 +36184,13 @@ void runloop_msg_queue_push(const char *msg,
       enum message_queue_icon icon,
       enum message_queue_category category)
 {
-   struct rarch_state *p_rarch = &rarch_st;
+   struct rarch_state *p_rarch                  = &rarch_st;
 #if defined(HAVE_GFX_WIDGETS)
-   bool widgets_active         = runloop_state.widgets_active;
+   bool widgets_active                          = runloop_state.widgets_active;
 #endif
 #ifdef HAVE_ACCESSIBILITY
-   settings_t *settings        = p_rarch->configuration_settings;
-   bool accessibility_enable   = settings->bools.accessibility_enable;
+   settings_t *settings                         = p_rarch->configuration_settings;
+   bool accessibility_enable                    = settings->bools.accessibility_enable;
    unsigned accessibility_narrator_speech_speed = settings->uints.accessibility_narrator_speech_speed;
 #endif
 
@@ -37044,73 +36308,71 @@ static bool input_driver_toggle_button_combo(
             return true;
          break;
       case INPUT_TOGGLE_HOLD_START:
-      {
-         static rarch_timer_t timer = {0};
-
-         if (!BIT256_GET_PTR(p_input, RETRO_DEVICE_ID_JOYPAD_START))
          {
-            /* timer only runs while start is held down */
-            RARCH_TIMER_END(timer);
-            return false;
+            static rarch_timer_t timer = {0};
+
+            if (!BIT256_GET_PTR(p_input, RETRO_DEVICE_ID_JOYPAD_START))
+            {
+               /* timer only runs while start is held down */
+               RARCH_TIMER_END(timer);
+               return false;
+            }
+
+            /* User started holding down the start button, start the timer */
+            if (!timer.timer_begin)
+            {
+               uint64_t current_usec = cpu_features_get_time_usec();
+               RARCH_TIMER_BEGIN_NEW_TIME_USEC(timer,
+                     current_usec,
+                     HOLD_BTN_DELAY_SEC * 1000000);
+               timer.timer_begin     = true;
+               timer.timer_end       = false;
+            }
+
+            RARCH_TIMER_TICK(timer, current_time);
+
+            if (!timer.timer_end && RARCH_TIMER_HAS_EXPIRED(timer))
+            {
+               /* start has been held down long enough,
+                * stop timer and enter menu */
+               RARCH_TIMER_END(timer);
+               return true;
+            }
          }
-
-         /* User started holding down the start button, start the timer */
-         if (!timer.timer_begin)
-         {
-            uint64_t current_usec = cpu_features_get_time_usec();
-            RARCH_TIMER_BEGIN_NEW_TIME_USEC(timer,
-                  current_usec,
-                  HOLD_BTN_DELAY_SEC * 1000000);
-            timer.timer_begin     = true;
-            timer.timer_end       = false;
-         }
-
-         RARCH_TIMER_TICK(timer, current_time);
-
-         if (!timer.timer_end && RARCH_TIMER_HAS_EXPIRED(timer))
-         {
-            /* start has been held down long enough,
-             * stop timer and enter menu */
-            RARCH_TIMER_END(timer);
-            return true;
-         }
-
-         return false;
-      }
+         break;
       case INPUT_TOGGLE_HOLD_SELECT:
-      {
-         static rarch_timer_t timer = {0};
-
-         if (!BIT256_GET_PTR(p_input, RETRO_DEVICE_ID_JOYPAD_SELECT))
          {
-            /* timer only runs while select is held down */
-            RARCH_TIMER_END(timer);
-            return false;
+            static rarch_timer_t timer = {0};
+
+            if (!BIT256_GET_PTR(p_input, RETRO_DEVICE_ID_JOYPAD_SELECT))
+            {
+               /* timer only runs while select is held down */
+               RARCH_TIMER_END(timer);
+               return false;
+            }
+
+            /* user started holding down the select button, start the timer */
+            if (!timer.timer_begin)
+            {
+               uint64_t current_usec = cpu_features_get_time_usec();
+               RARCH_TIMER_BEGIN_NEW_TIME_USEC(timer,
+                     current_usec,
+                     HOLD_BTN_DELAY_SEC * 1000000);
+               timer.timer_begin     = true;
+               timer.timer_end       = false;
+            }
+
+            RARCH_TIMER_TICK(timer, current_time);
+
+            if (!timer.timer_end && RARCH_TIMER_HAS_EXPIRED(timer))
+            {
+               /* select has been held down long enough,
+                * stop timer and enter menu */
+               RARCH_TIMER_END(timer);
+               return true;
+            }
          }
-
-         /* user started holding down the select button, start the timer */
-         if (!timer.timer_begin)
-         {
-            uint64_t current_usec = cpu_features_get_time_usec();
-            RARCH_TIMER_BEGIN_NEW_TIME_USEC(timer,
-                  current_usec,
-                  HOLD_BTN_DELAY_SEC * 1000000);
-            timer.timer_begin     = true;
-            timer.timer_end       = false;
-         }
-
-         RARCH_TIMER_TICK(timer, current_time);
-
-         if (!timer.timer_end && RARCH_TIMER_HAS_EXPIRED(timer))
-         {
-            /* select has been held down long enough,
-             * stop timer and enter menu */
-            RARCH_TIMER_END(timer);
-            return true;
-         }
-
-         return false;
-      }
+         break;
       default:
       case INPUT_TOGGLE_NONE:
          break;
@@ -37141,7 +36403,8 @@ static bool menu_display_libretro(
 
       core_run();
       runloop_state.libretro_core_runtime_usec        +=
-         rarch_core_runtime_tick(slowmotion_ratio, current_time);
+         runloop_core_runtime_tick(&runloop_state,
+               slowmotion_ratio, current_time);
       runloop_state.input_driver_block_libretro_input  = false;
 
       return false;
@@ -38576,7 +37839,8 @@ int runloop_iterate(void)
 
    /* Increment runtime tick counter after each call to
     * core_run() or run_ahead() */
-   runloop_state.libretro_core_runtime_usec += rarch_core_runtime_tick(
+   runloop_state.libretro_core_runtime_usec += runloop_core_runtime_tick(
+         &runloop_state,
          slowmotion_ratio,
          current_time);
