+++ retroarch.h
@@ -27,6 +27,10 @@
 #include <retro_inline.h>
 #include <retro_common_api.h>
 
+#ifdef HAVE_THREADS
+#include <rthreads/rthreads.h>
+#endif
+
 #ifdef HAVE_CONFIG_H
 #include "config.h"
 #endif
@@ -43,6 +47,7 @@
 
 #include "core_type.h"
 #include "core.h"
+#include "core_option_manager.h"
 
 #ifdef HAVE_MENU
 #include "menu/menu_defines.h"
@@ -78,6 +83,9 @@ RETRO_BEGIN_DECLS
                                             * 2 - Normal
                                             * 3 - Late
                                             */
+
+#define MEASURE_FRAME_TIME_SAMPLES_COUNT (2 * 1024)
+#define AUDIO_BUFFER_FREE_SAMPLES_COUNT  (8 * 1024)
 
 enum rarch_ctl_state
 {
@@ -320,6 +328,12 @@ bool rarch_ctl(enum rarch_ctl_state state, void *data);
 
 int retroarch_get_capabilities(enum rarch_capabilities type,
       char *s, size_t len);
+
+bool retroarch_validate_game_options(
+      char *s, size_t len, bool mkdir);
+
+bool retroarch_validate_folder_options(
+      char *s, size_t len, bool mkdir);
 
 void retroarch_override_setting_set(enum rarch_override_setting enum_idx, void *data);
 
@@ -2018,6 +2032,290 @@ typedef enum apple_view_type
    APPLE_VIEW_TYPE_VULKAN,
    APPLE_VIEW_TYPE_METAL
 } apple_view_type_t;
+
+typedef struct input_game_focus_state
+{
+   bool enabled;
+   bool core_requested;
+} input_game_focus_state_t;
+
+
+typedef struct runloop runloop_state_t;
+
+struct runloop
+{
+   double audio_source_ratio_original;
+   double audio_source_ratio_current;
+   struct retro_system_av_info av_info; /* double alignment */
+
+   retro_time_t libretro_core_runtime_last;
+   retro_time_t libretro_core_runtime_usec;
+   retro_time_t frame_limit_minimum_time;
+   retro_time_t frame_limit_last_time;
+   retro_time_t frame_time_samples[
+      MEASURE_FRAME_TIME_SAMPLES_COUNT];
+
+   retro_usec_t frame_time_last;        /* int64_t alignment */
+
+   uint64_t frame_time_count;
+   uint64_t frame_count;
+#ifdef HAVE_RUNAHEAD
+   uint64_t last_frame_count_runahead;
+#endif
+
+   uint64_t free_audio_samples_count;
+
+   float   *audio_output_samples_buf;
+   float   *audio_input_data;
+
+   int16_t *audio_output_samples_conv_buf;
+#ifdef HAVE_REWIND
+   int16_t *audio_rewind_buf;
+#endif
+
+   struct retro_audio_callback audio_callback;   /* ptr alignment */
+   msg_queue_t msg_queue;                        /* ptr alignment */
+#ifdef HAVE_THREADS
+   slock_t *msg_queue_lock;
+   slock_t *display_lock;
+   slock_t *context_lock;
+#endif
+
+   core_option_manager_t *core_options;
+   retro_keyboard_event_t key_event;             /* ptr alignment */
+   retro_keyboard_event_t frontend_key_event;    /* ptr alignment */
+
+   rarch_system_info_t system;                   /* ptr alignment */
+   struct retro_frame_time_callback frame_time;  /* ptr alignment */
+   struct retro_audio_buffer_status_callback audio_buffer_status; /* ptr alignment */
+
+   void *audio_context_audio_data;
+   void *audio_resampler_data;
+
+#ifdef HAVE_REWIND
+   size_t audio_rewind_ptr;
+   size_t audio_rewind_size;
+#endif
+#ifdef HAVE_RUNAHEAD
+   size_t runahead_save_state_size;
+#endif
+   size_t audio_buffer_size;
+   size_t audio_data_ptr;
+   size_t audio_chunk_size;
+   size_t audio_chunk_nonblock_size;
+   size_t audio_chunk_block_size;
+   size_t msg_queue_size;
+
+   unsigned pending_windowed_scale;
+   unsigned max_frames;
+   unsigned audio_latency;
+   unsigned free_audio_samples_buf[
+      AUDIO_BUFFER_FREE_SAMPLES_COUNT];
+
+   struct retro_fastforwarding_override fastmotion_override; /* float alignment */
+   float audio_rate_control_delta;
+   float audio_input_sample_rate;
+   float audio_volume_gain;
+#ifdef HAVE_AUDIOMIXER
+   float audio_mixer_volume_gain;
+#endif
+
+   input_game_focus_state_t game_focus_state; /* bool alignment */
+#ifdef HAVE_GFX_WIDGETS
+   bool widgets_active;
+   bool widgets_persisting;
+   bool widgets_paused;
+   bool widgets_fast_forward;
+   bool widgets_rewinding;
+#endif
+   bool audio_active;
+   bool audio_use_float;
+   bool audio_suspended;
+   bool audio_control;
+   bool audio_mute_enable;
+#ifdef HAVE_AUDIOMIXER
+   bool audio_mixer_mute_enable;
+   bool audio_mixer_active;
+#endif
+   bool missing_bios;
+   bool force_nonblock;
+   bool paused;
+   bool idle;
+   bool slowmotion;
+   bool fastmotion;
+   bool shutdown_initiated;
+   bool core_shutdown_initiated;
+   bool core_running;
+   bool perfcnt_enable;
+   bool video_active;
+   bool video_started_fullscreen;
+#ifdef HAVE_RUNAHEAD
+   bool runahead_video_active;
+#endif
+   bool game_options_active;
+   bool folder_options_active;
+   bool autosave;
+#ifdef HAVE_CONFIGFILE
+   bool overrides_active;
+   bool remaps_core_active;
+   bool remaps_game_active;
+   bool remaps_content_dir_active;
+#endif
+#ifdef HAVE_SCREENSHOTS
+   bool max_frames_screenshot;
+   char max_frames_screenshot_path[PATH_MAX_LENGTH];
+#endif
+#ifdef HAVE_NETWORKING
+/* Only used before init_netplay */
+   bool netplay_enabled;
+   bool netplay_is_client;
+   /* Used to avoid recursive netplay calls */
+   bool in_netplay;
+   bool netplay_client_deferred;
+   bool is_mitm;
+#endif
+   bool has_set_username;
+   bool rarch_error_on_init;
+   bool rarch_force_fullscreen;
+   bool has_set_core;
+   bool has_set_verbosity;
+   bool has_set_libretro;
+   bool has_set_libretro_directory;
+   bool has_set_save_path;
+   bool has_set_state_path;
+#ifdef HAVE_PATCH
+   bool has_set_ups_pref;
+   bool has_set_bps_pref;
+   bool has_set_ips_pref;
+#endif
+#ifdef HAVE_QT
+   bool qt_is_inited;
+#endif
+   bool has_set_log_to_file;
+   bool rarch_is_inited;
+   bool rarch_is_switching_display_mode;
+   bool rarch_is_sram_load_disabled;
+   bool rarch_is_sram_save_disabled;
+   bool rarch_use_sram;
+   bool rarch_ups_pref;
+   bool rarch_bps_pref;
+   bool rarch_ips_pref;
+#ifdef HAVE_PATCH
+   bool rarch_patch_blocked;
+#endif
+   bool video_driver_window_title_update;
+
+   /**
+    * dynamic.c:dynamic_request_hw_context will try to set
+    * flag data when the context
+    * is in the middle of being rebuilt; in these cases we will save flag
+    * data and set this to true.
+    * When the context is reinit, it checks this, reads from
+    * deferred_flag_data and cleans it.
+    *
+    * TODO - Dirty hack, fix it better
+    */
+   bool deferred_video_context_driver_set_flags;
+   bool ignore_environment_cb;
+   bool core_set_shared_context;
+
+   /* Graphics driver requires RGBA byte order data (ABGR on little-endian)
+    * for 32-bit.
+    * This takes effect for overlay and shader cores that wants to load
+    * data into graphics driver. Kinda hackish to place it here, it is only
+    * used for GLES.
+    * TODO: Refactor this better. */
+   bool video_driver_use_rgba;
+
+   /* If set during context deinit, the driver should keep
+    * graphics context alive to avoid having to reset all
+    * context state. */
+   bool video_driver_cache_context;
+
+   /* Set to true by driver if context caching succeeded. */
+   bool video_driver_cache_context_ack;
+
+#ifdef HAVE_ACCESSIBILITY
+   /* Is text-to-speech accessibility turned on? */
+   bool accessibility_enabled;
+#endif
+#ifdef HAVE_CONFIGFILE
+   bool rarch_block_config_read;
+#endif
+#if defined(HAVE_CG) || defined(HAVE_GLSL) || defined(HAVE_SLANG) || defined(HAVE_HLSL)
+   bool cli_shader_disable;
+#endif
+
+   bool location_driver_active;
+   bool bluetooth_driver_active;
+   bool wifi_driver_active;
+   bool camera_driver_active;
+#ifdef HAVE_VIDEO_FILTER
+   bool video_driver_state_out_rgb32;
+#endif
+   bool video_driver_crt_switching_active;
+   bool video_driver_crt_dynamic_super_width;
+   bool video_driver_threaded;
+
+#ifdef HAVE_RUNAHEAD
+   bool has_variable_update;
+   bool runahead_save_state_size_known;
+   bool request_fast_savestate;
+   bool hard_disable_audio;
+
+   bool input_is_dirty;
+#endif
+
+#if defined(HAVE_NETWORKING)
+   bool has_set_netplay_mode;
+   bool has_set_netplay_ip_address;
+   bool has_set_netplay_ip_port;
+   bool has_set_netplay_stateless_mode;
+   bool has_set_netplay_check_frames;
+#endif
+
+   bool input_driver_keyboard_linefeed_enable;
+
+   bool input_driver_block_hotkey;
+   bool input_driver_block_libretro_input;
+   bool input_driver_nonblock_state;
+   bool input_driver_grab_mouse_state;
+
+#ifdef HAVE_MENU
+   bool menu_input_dialog_keyboard_display;
+   /* Is the menu driver still running? */
+   bool menu_driver_alive;
+   /* Are we binding a button inside the menu? */
+   bool menu_driver_is_binding;
+#endif
+
+   bool recording_enable;
+   bool streaming_enable;
+
+   bool midi_drv_input_enabled;
+   bool midi_drv_output_enabled;
+
+   bool midi_drv_output_pending;
+
+   bool main_ui_companion_is_on_foreground;
+   bool keyboard_mapping_blocked;
+#if defined(HAVE_CG) || defined(HAVE_GLSL) || defined(HAVE_SLANG) || defined(HAVE_HLSL)
+   bool shader_presets_need_reload;
+#endif
+#ifdef HAVE_RUNAHEAD
+   bool runahead_available;
+   bool runahead_secondary_core_available;
+   bool runahead_force_input_dirty;
+#endif
+};
+
+bool retroarch_validate_per_core_options(char *s,
+      size_t len, bool mkdir,
+      const char *core_name, const char *game_name);
+
+void core_option_manager_flush(
+      config_file_t *conf,
+      core_option_manager_t *opt);
 
 RETRO_END_DECLS
