+++ retroarch_fwd_decls.h
@@ -44,14 +44,18 @@ static bool midi_driver_set_all_sounds_off(struct rarch_state *p_rarch);
 static const void *midi_driver_find_handle(int index);
 static bool midi_driver_flush(void);
 
-static void runloop_deinit_core_options(
+retro_time_t runloop_core_runtime_tick(
+      runloop_state_t *p_runloop,
+      float slowmotion_ratio,
+      retro_time_t current_time);
+void runloop_deinit_core_options(
       runloop_state_t *p_runloop,
       const char *path_core_options);
-static core_option_manager_t *runloop_init_core_variables(
+core_option_manager_t *runloop_init_core_variables(
       runloop_state_t *p_runloop,
       settings_t *settings,
       const struct retro_variable *vars);
-static core_option_manager_t *runloop_init_core_options(
+core_option_manager_t *runloop_init_core_options(
       runloop_state_t *p_runloop,
       settings_t *settings,
       const struct retro_core_option_definition *option_defs);
